// src/enums.ts
export enum TrackStatus {
  PENDING_REVIEW = "PENDING_REVIEW",
  READY_TO_REVIEW = "READY_TO_REVIEW",
  REVIEWED = "REVIEWED",
  REVIEW_IN_PROGRESS = "REVIEW_IN_PROGRESS",
  READY_TO_SEND = "READY_TO_SEND",
  SEND = "SEND",
}

export enum InviteStatus {
  INVITED = "INVITED",
  ACCEPTED = "ACCEPTED",
  DECLINED = "DECLINED",
}